requirements.txt
streamlit
pandas
pytesseract
opencv-python
pillow